+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # Changelog
 
+## v1.6.0 (August 25, 2021)
+
+### Added
+
+-   New icons: Output Flow, Input Flow, Smoke Detector, File PDF, File DOC, Load Factor.
+
+### Changed
+
+-   Modified the style of Air Filter, Power Factor.
+
 ## v1.5.0 (August 17, 2021)
 
-### Added:
+### Added
 
 -   Device- and channel-related icons: ATS, ATS Alt, Distribution Box, Distribution Box Outline, Drive, Meter Face, Motor Alt, Motor Controller Outline, Motor Controller, Motor Outline, Motor Starter, Motor, Relay, Sensor Alt, Sensor Device Alt, Sensor Device, Sensor, Device Info, Trip Curve Alt, Trip Curve, Trip Unit, UPS Outline, UPS, Configuration, Device Configuration, Control, Diagnostic Alt, Diagnostic, Apparent Energy, Apparent Power, Bypass Alt, Bypass, Flood, Frequency, Harmonics, Phase Angle, Power Factor, Reactive Energy, Reactive Power, Real Energy, Real Power, Sequence, Air Conditioner, Fan Circled.